Handle more cases in resgrep
before, resgrep assumed res was a subdir of our current directory...

now, searching from a subdirectory of a res directory (e.g. within a specific layout directory you're cded into) is handled

additionally, if you're in a subdirectory that has a parent that contains a res directory (as would be the case if you were in a packages src directory), that entire package's res directory is searched.

ps2: Fix "child of res" case

Change-Id: Iaadbb37d399bfb41a50ea63e492153e29cd2139d