Colorize javac errors
You know those multi-line ones that are frequently interspersed with other buld output?
The ones that don't even have the word "error" in them to search for?

This makes them red.

ps2: ensure the temporary stderr file gets killed if no errors occur

The build performance impact is negligible in comparison to vanir's mkc, which uses sed
to parse every line to highlight errors and warnings from make and gcc output.
This, on the other hand, only uses i/o redirection, and a quick conditional cat of a file
containing javac's stderr if it finished with a non-zero exit code and the file exists.

Change-Id: I3b9b7d8a0c76958588ac1603b6742987d6dde54c
Signed-off-by: nuclearmistake <nuclearmistake@gmail.com>