build: Add chromium prebuilt support to envsetup.sh && The core Makefile
Squashed commit:
build: Fix PreBuilt Chromium refs
https://github.com/camcory/platform_build/commit/03a8bff75be03a93ed5ceb9fa5e8eb74d5e090b2

This adds a chromium_prebuilt function to envsetup.sh that is invoked by lunch to check
whether the chromium prebuilts are up-to-date or not. If not, it will be built from source
and then the new source built version will be pulled during brunch/mka bacon to become the
new prebuilts for future builds.

This is all opt-in through the USE_PREBUILT_CHROMIUM flag. Without it being set to 1,
none of this would be ran, and regular operations will go on.

PS13:
-use export TARGET_DEVICE
-replace git -C with params compatible to old git versions

Change-Id: I40f8c8dbd2a8a84a5c1b9f47ee04180a71ef4e07