ASAN: set MAXIMUM_OVERDRIVE := true to disable address sanitizer
with respect to BONE_STOCK taking priority.  this reduced my build time by ~20%.

Address sanitizer is a debugging tool that detects memory errors in the program through debugging tools
like gdb. It does so by using an extra instrumentation module llvm pass during compilation and a runtime
library that replaces the malloc function and by inserting frame pointers in the stack.

Output code performance can be effected if a "hot" function that is known to speed up the code gets
flagged during the extra compiler pass but the primary benefit of turning it off is to speed up your
build time from clobber.

Change-Id: I7cdd2c10e37255b5f5b31f5999b6a9d530eb9493