BUILD: add rule kernel to create flashable zip of TARGET_PRODUCT kernel
Create build rule for 'mka kernel' to create a flashable kernel zip for
the lunched build named in /out/target/product/devicename. This requires
a subtree of Koush's AnyKernel tool to be created.

PS2: versioning
PS3: clean up .conf and autogen files
PS4: sign zip with security certs
PS5: clone into a subtree of AnyKernel if the tool is not present

Change-Id: I962b75f51a12d921d82d059e97bd0ed39eea0b80