Do not fail build if common.Cleanup does not find and/or remove temporary file /tmp/system-*.img
Something is removing it in some builds, prior to common.Cleanup(i)

I think it was introduced from this commit...
https://github.com/VanirAOSP/build/commit/8a2aadc5164d0a112c09ff6c57aa5bfcbf509bf1

I have not found where/why it is being removed yet. It does not occur
everytime. It happens when I run my build script only on the first
device's build, and the rest work just fine.

For now just allow an exception so that the entire build does not fail.

Change-Id: I963e753db6aad9ebab6050e8d2241524d07638e9
Signed-off-by: sudosurootdev <sudosurootdev@gmail.com>