Do not fail entire build if common.Cleanup does not find temporary file /tmp/system-*.img to remove. Something is removing it in some cases. I think it was introduced from this commit...
https://github.com/VanirAOSP/build/commit/8a2aadc5164d0a112c09ff6c57aa5bfcbf509bf1

"releasetools: Add bootable image signing"...

I am not positive, and I have not found where it is being removed yet.
It does not occur everytime either. It happens when I run my build
script on the first device's build, and the rest work fine. For now
just allow an exception so that the entire build is not failed.

Change-Id: If70292597985d89d2304688a0ec06e8c812e9e35
Signed-off-by: sudosurootdev <sudosurootdev@gmail.com>