add -flto-compression-level flag to lto flags list
-flto-compression-level=n
This option specifies the level of compression used for intermediate language written to LTO object files, 
and is only meaningful in conjunction with LTO mode (-fwhopr, -flto).
Valid values are 0 (no compression) to 9 (maximum compression). Values outside this range are clamped to either 0 or 9. 
If the option is not given, a default balanced compression setting is used. 

Change-Id: I9f4ad928b43f9b2e3149f182a175f8ba29728273